refactor: 이미지 파일명 정규화
