+++ src/screens/Map/Map.tsx
@@ -5,8 +5,8 @@ import NaverMapView, { Marker } from 'react-native-nmap';
 import ListIcon from 'src/assets/icons/icon_list.svg';
 import MoreIcon from 'src/assets/icons/icon_more.svg';
 import PersonWhiteIcon from 'src/assets/icons/icon_person_white.svg';
-import mapPickerImage from 'src/assets/images/icon_mappicker.png';
-import mapPickerSelectImage from 'src/assets/images/icon_mappicker_select.png';
+import mapPickerImage from 'src/assets/images/image_mappicker.png';
+import mapPickerSelectImage from 'src/assets/images/image_mappicker_select.png';
 import Header from 'src/components/Header/Header';
 import Typo from 'src/components/Typo/Typo';
 import useGeolocation from 'src/hooks/useGeolocation';
